WIP clarifying WhereZvm
