+++ webcomponents/src/viewModels/where.perspective.ts
@@ -1,17 +1,33 @@
 import {AgentPubKeyB64, ActionHashB64, EntryHashB64, Dictionary} from "@holochain-open-dev/core-types";
 import {HereEntry, HereOutput} from "./where.bindings";
-import {MarkerType} from "./playset.perspective";
-import {EmojiGroupVariant, MarkerPiece, SpaceEntry, SvgMarkerVariant} from "./playset.bindings";
-import {HoloHashedB64, mapReplacer, mapReviver} from "../utils";
-import {PlaysetEntry} from "./ludotheque.bindings";
-
+import {MarkerType, Space} from "./playset.perspective";
+import {mapReplacer, mapReviver} from "../utils";
 
 
 /** */
 export interface WherePerspective {
   plays: Dictionary<Play>,
-  sessions: Dictionary<EntryHashB64[]>,
-  currentSessions: Dictionary<EntryHashB64>,
+  /* SessionEh -> [locations] */
+  //locations: Dictionary<Location[]>,
+  /* SpaceEh -> [sessions] */
+  //sessions: Dictionary<PlacementSession[]>,
+}
+
+
+/** A 'Play' is a live 'Space' with locations and sessions */
+export interface Play {
+  space: Space,
+  visible: boolean;
+  sessions: Dictionary<PlacementSession>,
+}
+
+
+
+/** */
+export interface PlacementSession {
+  name: string,
+  index: number,
+  locations: (LocationInfo | null)[];
 }
 
 
@@ -57,90 +73,8 @@ export interface Coord {
 }
 
 
-export interface Space {
-  name: string;
-  origin: EntryHashB64;
-  surface: any;
-  maybeMarkerPiece?: MarkerPiece;
-  meta: PlayMeta;
-}
 
-
-
-export interface PlacementSession {
-  name: string,
-  index: number,
-  locations: (LocationInfo | null)[];
-}
-
-/** A 'Play' is a live 'Space' with locations and sessions */
-
-export interface Play {
-  space: Space,
-  visible: boolean;
-  sessions: Dictionary<PlacementSession>,
-}
-
-
-export interface PlayMeta {
-  ui: UiItem[],
-  subMap: Map<string, string>,
-  /* Marker */
-  markerType: MarkerType,
-  singleEmoji: string,
-  //emojiGroup: EntryHashB64 | null,
-  //svgMarker: EntryHashB64 | null,
-  /* Tag */
-  multi: boolean,
-  canTag: boolean,
-  tagVisible: boolean,
-  tagAsMarker: boolean,
-  predefinedTags: string[],
-  /* Session */
-  sessionCount: number,
-  canModifyPast: boolean,
-  sessionLabels: string[],
-}
-
-
-
-export interface UiItem {
-  box: UiBox,
-  style: string,
-  content: string,
-}
-
-
-export interface UiBox {
-  width: number,
-  height: number,
-  left: number,
-  top: number,
-}
-
-
-export function defaultPlayMeta(): PlayMeta {
-  return  {
-    subMap: new Map(),
-    ui: [],
-    /* Marker */
-    markerType: MarkerType.Avatar,
-    multi: false,
-    singleEmoji: "😀",
-    //emojiGroup: null,
-    //svgMarker: null,
-    /* Tag */
-    canTag: false,
-    tagVisible: false,
-    tagAsMarker: false,
-    predefinedTags: [],
-    /* Sessions */
-    sessionCount: 2,
-    canModifyPast: true,
-    sessionLabels: [],
-  } as PlayMeta
-}
-
+/** */
 export function defaultLocationMeta(): LocationMeta {
   return  {
     markerType: MarkerType.Tag,
@@ -150,59 +84,11 @@ export function defaultLocationMeta(): LocationMeta {
     authorName: "",
     emoji: "",
     svgMarker: "",
-  } as LocationMeta
+  }
 }
 
 
 /** -- Conversions -- */
-
-
-/** */
-export function convertEntryToSpace(entry: SpaceEntry): Space {
-  return {
-    name: entry.name,
-    origin: entry.origin,
-    surface: JSON.parse(entry.surface),
-    maybeMarkerPiece: entry.maybeMarkerPiece,
-    meta: entry.meta ? convertEntryToPlayMeta(entry.meta) : defaultPlayMeta(),
-  }
-}
-
-/** */
-export function convertSpaceToEntry(space: Space): SpaceEntry {
-  return {
-    name: space.name,
-    origin: space.origin,
-    surface: JSON.stringify(space.surface),
-    maybeMarkerPiece: space.maybeMarkerPiece,
-    meta: convertPlayMetaToEntryField(space.meta)
-  }
-}
-
-/** */
-export function convertEntryToPlayMeta(meta: Dictionary<string>): PlayMeta {
-  let spaceMeta: any = {};
-  try {
-    for (const [key, value] of Object.entries(meta)) {
-      Object.assign(spaceMeta, {[key]: JSON.parse(value, mapReviver)})
-    }
-  } catch (e) {
-    console.error("Failed parsing meta filed into PlayMeta")
-    console.error(e)
-  }
-  //console.log({spaceMeta})
-  return spaceMeta as PlayMeta;
-}
-
-/** */
-export function convertPlayMetaToEntryField(playMeta: PlayMeta): Dictionary<string> {
-  let dic: Dictionary<string> = {};
-  for (const [key, value] of Object.entries(playMeta)) {
-    dic[key] = JSON.stringify(value, mapReplacer)
-  }
-  //console.log({dic})
-  return dic
-}
 
 /** */
 export function convertHereToLocation(info: HereInfo) : LocationInfo {
@@ -241,50 +127,3 @@ export function convertLocationToHere(location: WhereLocation) : HereEntry {
   } as HereEntry
 }
 
-
-/** */
-export function createPlayset(name: string, spaces: HoloHashedB64<SpaceEntry>[]): PlaysetEntry {
-  console.log("newPlayset() called:")
-  console.log({spaces})
-  /* Get templates */
-  let templates = new Array();
-  for (const space of spaces) {
-    if (!templates.includes(space.content.origin)) {
-      templates.push(space.content.origin)
-    }
-  }
-  /* Get markers */
-  let svgMarkers = new Array();
-  let emojiGroups = new Array();
-  for (const entry of spaces) {
-    let space = convertEntryToSpace(entry.content);
-    if (space.meta.markerType == MarkerType.SvgMarker) {
-      let markerEh = (space.maybeMarkerPiece! as SvgMarkerVariant).svg;
-      if (markerEh && !svgMarkers.includes(markerEh)) {
-        svgMarkers.push(markerEh)
-      }
-    } else {
-      if (space.meta.markerType == MarkerType.EmojiGroup) {
-        let eh = (space.maybeMarkerPiece! as EmojiGroupVariant).emojiGroup;
-        if (eh && !svgMarkers.includes(eh)) {
-          emojiGroups.push(eh)
-        }
-      }
-    }
-
-  }
-  /* Get space hashes */
-  let spaceEhs = new Array();
-  for (const space of spaces) {
-    spaceEhs.push(space.hash)
-  }
-  /* - Create PlaysetEntry */
-  return {
-    name,
-    description: "",
-    spaces: spaceEhs,
-    templates,
-    svgMarkers,
-    emojiGroups,
-  } as PlaysetEntry;
-}
